% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_col_cube.R
\name{geom_col_cube}
\alias{geom_col_cube}
\title{Column plot with rounded tops}
\usage{
geom_col_cube(
  radius = ggplot2::unit(0.5, "npc"),
  position = ggplot2::position_dodge(),
  ...
)
}
\arguments{
\item{radius}{The radius of the rounded corners, given as a unit object.
This defaults to \code{grid::unit(.5, "npc")}, where \code{"npc"} is "Normalised Parent Coordinates".
This effectively make the radius half the width of the column (the parent).}

\item{position}{The position is set to \code{position_dodge()} so colours are positioned side by
side. The results of \code{\link[ggrounded:geom_col_rounded]{ggrounded::geom_col_rounded()}} with the default \code{position_stack}
are aesthetically questionable.}

\item{...}{Other arguments passed on to \code{\link[ggrounded:geom_col_rounded]{ggrounded::geom_col_rounded()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
A ggplot object.
}
\description{
Create a column plot with rounded tops using
'ggplot2'. \code{geom_col_cube()} is a wrapper function utilising
the \code{\link[ggrounded:geom_col_rounded]{ggrounded::geom_col_rounded()}} to produce column charts that match
the Cube Group Brand Guidelines.

This is achieved by setting defaults for two parameters \code{radius} and \code{position}
}
\examples{
library(ggplot2)

ggplot(data.frame(x = letters[1:3], y = c(2.3, 1.9, 3.2)), aes(x, y)) +
  geom_col_cube()

}
