% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_line_cube.R
\name{geom_line_cube}
\alias{geom_line_cube}
\title{Line plots in the Cube Group style rounded tops}
\usage{
geom_line_cube(lineend = "round", linewidth = 1.2, ...)
}
\arguments{
\item{lineend}{defaults to \code{"round"}to match the brand guidelines.}

\item{linewidth}{defaults to 1.2 to produce generally clearer charts.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} . These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} . They may also be parameters to the paired geom/stat.}
}
\value{
A ggplot object.
}
\description{
Create a line plot with rounded ends using
'ggplot2'. \code{geom_line_cube()} is a wrapper function utilising
the \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} with defaults that produce line charts that match
the Cube Group Brand Guidelines.

This is achieved by setting defaults for two parameters \code{lineend} and \code{linewidth}
}
\examples{
library(ggplot2)

# To be added

}
