% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cube_fonts.R
\name{install_cube_fonts}
\alias{install_cube_fonts}
\title{Install Cube Fonts}
\usage{
install_cube_fonts(...)
}
\description{
Themes in this package use the Google Open Sans font as an alternative to Agenda.
This is required to enable easy loading of the package and deployment.
}
\details{
This function allows these to be installed within the package
}
\seealso{
Other cubethemes: 
\code{\link{theme_cube_foundation}()},
\code{\link{theme_cube_map}()}
}
\concept{cubethemes}
