% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_indicator.R
\name{recode_indicator}
\alias{recode_indicator}
\title{A simple function for recoding qualitative variables.}
\usage{
recode_indicator(x)
}
\arguments{
\item{x}{the indicator variable to be recoded.}
}
\value{
a numeric value 1 or 0
}
\description{
Use this function to recode Qualitative Yes/No data to 1/0.
Note this version converts all non-Yes answers to 0
}
\examples{
indicators = c("Yes","No","Unsure","y","No")
recode_indicator(indicators)

# To be added# Define function to recode indicators
}
