% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_cube.R
\name{scale_colour_cube}
\alias{scale_colour_cube}
\title{ggplots in the Cube Group palette}
\usage{
scale_colour_cube(cube_theme = c("green", "orange", "white", "grey"))
}
\arguments{
\item{cube_theme}{defaults to \code{"white"}to select colours that work on a white background}

\item{...}{Other arguments passed on to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot object.
}
\description{
Use the cube palette in 'ggplot2'.
\code{scale_colour_cube()} is a wrapper function utilising
the \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} that utilises a switch function to select a palette
that works with the cube themes.
}
\examples{
library(ggplot2)

# To be added

}
