% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_green.R
\name{theme_cube_green}
\alias{theme_cube_green}
\title{Cube Green Theme}
\usage{
theme_cube_green(
  base.colour = cube.darkgreen,
  base.size = 10,
  line.colour = cube.pink,
  text.colour = cube.pink
)
}
\description{
This theme takes \code{\link[cubieR]{theme_cube_foundation}()} and sets
background colour to green and line and text colours to pink.
}
\details{
Colours are sourced from \code{cubepalette} and are Cube Brand Guidelines compliant.
}
\seealso{
Other cubieR: 
\code{\link{theme_cube_grey}()},
\code{\link{theme_cube_orange}()},
\code{\link{theme_cube_pink}()},
\code{\link{theme_cube_white}()}
}
\concept{cubieR}
