% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_map.R
\name{theme_cube_map}
\alias{theme_cube_map}
\title{Cube Map Theme}
\usage{
theme_cube_map(
  base_colour = c("white", "green", "orange", "grey"),
  base_size = 10,
  base_family = "Open Sans"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is designed to be a foundation from which to build new
themes, and not meant to be used directly. \code{theme_cube_map()}
is a complete theme with only minimal number of elements defined.
It is easier to create new Cube themes by extending this one rather
than \code{\link[ggplot2]{theme_gray}()} or \code{\link[ggplot2]{theme_bw}()},
because those themes define elements deep in the hierarchy.
}
\details{
This theme takes \code{\link[ggthemes]{theme_map}()} and sets
base text family and size to \code{Agenda} and \code{15}.
}
\seealso{
Other cubethemes: 
\code{\link{install_cube_fonts}()},
\code{\link{theme_cube_foundation}()}
}
\concept{cubethemes}
