% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_pink.R
\name{theme_cube_pink}
\alias{theme_cube_pink}
\title{Cube Pink Theme}
\usage{
theme_cube_pink()
}
\description{
This theme is deprecated. Please use theme_cube_orange.
}
\seealso{
Other cubieR: 
\code{\link{theme_cube_green}()},
\code{\link{theme_cube_grey}()},
\code{\link{theme_cube_orange}()},
\code{\link{theme_cube_white}()}
}
\concept{cubieR}
