% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_white.R
\name{theme_cube_white}
\alias{theme_cube_white}
\title{Cube White Theme}
\usage{
theme_cube_white(
  base.colour = cube.white,
  base.size = 10,
  line.colour = cube.darkgreen,
  text.colour = cube.darkgreen
)
}
\description{
This theme takes \code{\link[cubieR]{theme_cube_foundation}()} and sets
background colour to white and line and text colours to green.
}
\details{
Colours are sourced from \code{cubepalette} and are Cube Brand Guidelines compliant.
}
\seealso{
Other cubieR: 
\code{\link{theme_cube_green}()},
\code{\link{theme_cube_grey}()},
\code{\link{theme_cube_orange}()},
\code{\link{theme_cube_pink}()}
}
\concept{cubieR}
